create Modal component and use it on PR settings
